Clean Warnings (?)

+++ src/core/core.cpp
@@ -106,7 +106,7 @@ System::ResultStatus System::Load(EmuWindow* emu_window, const std::string& file
 
     const Loader::ResultStatus load_result{app_loader->Load(current_process)};
     if (Loader::ResultStatus::Success != load_result) {
-        LOG_CRITICAL(Core, "Failed to load ROM (Error %i)!", load_result);
+        LOG_CRITICAL(Core, "Failed to load ROM (Error %i)!", static_cast<int>(load_result));
         System::Shutdown();
 
         switch (load_result) {
