+++ src/core/hle/service/service.cpp
@@ -149,7 +149,7 @@ ResultCode ServiceFrameworkBase::HandleSyncRequest(Kernel::HLERequestContext& co
         break;
     }
     default:
-        UNIMPLEMENTED_MSG("command_type=%d", context.GetCommandType());
+        UNIMPLEMENTED_MSG("command_type=%d", static_cast<int>(context.GetCommandType()));
     }
 
     context.WriteToOutgoingCommandBuffer(*Kernel::GetCurrentThread());
